requirements.txt
certifi==2023.7.22
charset-normalizer==3.1.0
idna==3.4
mysql-connector==2.2.9
Pillow==9.5.0
requests==2.31.0
urllib3==2.0.3
python-dotenv==1.0.0